pyproject.toml
[tool.poetry]
name = "emotion"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
selenium = "^4.27.1"
pandas = "^2.2.3"
psutil = "^6.1.0"
openpyxl = "^3.1.5"
flask = "^3.1.0"
requests = "^2.32.3"
google = "^3.0.0"
google-generativeai = "^0.8.3"
python-dotenv = "^1.0.1"
line-bot-sdk = "^3.14.2"
langchain = "^0.3.12"
openai = "^1.57.4"
langchain-community = "^0.3.12"
langchain-openai = "^0.2.12"
qdrant-client = "^1.12.2"
matplotlib = "^3.10.0"
scikit-learn = "^1.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
